Write new logic for ExchangeModel
In order to realise simplier structure of the model now all Exchanges (iso, aniso, dmi) and distance support the same processing format for reading and filtering. Main data now are stored in self.data dictionary. If you want to add new exchange or smth it is enough to add entries to self.keys, self.names_to_read, self.format_functions and self.update_attributes and to define format function for your entry. All reading and filtering functionality will be rady after that without modification of specific functions responsible for reading and filtering. NOTE: if you modifying the data is is important to call update_sttributes() at the end.